Add new completion item kinds
Split up more modules so we can compile
