+++ haskell-lsp-types/src/Language/Haskell/LSP/TH/DataTypesJSON.hs
@@ -14,13 +14,15 @@ module Language.Haskell.LSP.TH.DataTypesJSON
     ( module Language.Haskell.LSP.TH.DataTypesJSON
     , module Language.Haskell.LSP.TH.CodeAction
     , module Language.Haskell.LSP.TH.Command
+    , module Language.Haskell.LSP.TH.Completion
     , module Language.Haskell.LSP.TH.Diagnostic
+    , module Language.Haskell.LSP.TH.DocumentFilter
     , module Language.Haskell.LSP.TH.List
     , module Language.Haskell.LSP.TH.Location
     , module Language.Haskell.LSP.TH.MarkupContent
     , module Language.Haskell.LSP.TH.Message
     , module Language.Haskell.LSP.TH.Symbol
-    , module Language.Haskell.LSP.TH.TextDocumentIdentifier
+    , module Language.Haskell.LSP.TH.TextDocument
     , module Language.Haskell.LSP.TH.Uri
     , module Language.Haskell.LSP.TH.WorkspaceEdit
     ) where
@@ -36,141 +38,19 @@ import qualified Data.Text                                  as T
 import           Language.Haskell.LSP.TH.ClientCapabilities
 import           Language.Haskell.LSP.TH.CodeAction
 import           Language.Haskell.LSP.TH.Command
+import           Language.Haskell.LSP.TH.Completion
 import           Language.Haskell.LSP.TH.Constants
 import           Language.Haskell.LSP.TH.Diagnostic
+import           Language.Haskell.LSP.TH.DocumentFilter
 import           Language.Haskell.LSP.TH.List
 import           Language.Haskell.LSP.TH.Location
+import           Language.Haskell.LSP.TH.MarkupContent
 import           Language.Haskell.LSP.TH.Message
 import           Language.Haskell.LSP.TH.Symbol
-import           Language.Haskell.LSP.TH.TextDocumentIdentifier
+import           Language.Haskell.LSP.TH.TextDocument
 import           Language.Haskell.LSP.TH.Utils
 import           Language.Haskell.LSP.TH.Uri
 import           Language.Haskell.LSP.TH.WorkspaceEdit
-
--- ---------------------------------------------------------------------
-
-
-{-
-TextDocumentItem
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentitem
-
-    New: An item to transfer a text document from the client to the server.
-
-interface TextDocumentItem {
-    /**
-     * The text document's URI.
-     */
-    uri: string;
-
-    /**
-     * The text document's language identifier.
-     */
-    languageId: string;
-
-    /**
-     * The version number of this document (it will strictly increase after each
-     * change, including undo/redo).
-     */
-    version: number;
-
-    /**
-     * The content of the opened text document.
-     */
-    text: string;
-}
--}
-
-data TextDocumentItem =
-  TextDocumentItem {
-    _uri        :: Uri
-  , _languageId :: Text
-  , _version    :: Int
-  , _text       :: Text
-  } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''TextDocumentItem
-makeFieldsNoPrefix ''TextDocumentItem
-
--- ---------------------------------------------------------------------
-{-
-TextDocumentPositionParams
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentpositionparams
-
-    Changed: Was TextDocumentPosition in 1.0 with inlined parameters
-
-
-interface TextDocumentPositionParams {
-    /**
-     * The text document.
-     */
-    textDocument: TextDocumentIdentifier;
-
-    /**
-     * The position inside the text document.
-     */
-    position: Position;
-}
-
--}
-data TextDocumentPositionParams =
-  TextDocumentPositionParams
-    { _textDocument :: TextDocumentIdentifier
-    , _position     :: Position
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''TextDocumentPositionParams
-makeFieldsNoPrefix ''TextDocumentPositionParams
-
--- ---------------------------------------------------------------------
-{-
-New in 3.0
-----------
-
-DocumentFilter
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#new-documentfilter
-
-A document filter denotes a document through properties like language, schema or
-pattern. Examples are a filter that applies to TypeScript files on disk or a
-filter the applies to JSON files with name package.json:
-
-    { language: 'typescript', scheme: 'file' }
-    { language: 'json', pattern: '**/package.json' }
-
-export interface DocumentFilter {
-        /**
-         * A language id, like `typescript`.
-         */
-        language?: string;
-
-        /**
-         * A Uri [scheme](#Uri.scheme), like `file` or `untitled`.
-         */
-        scheme?: string;
-
-        /**
-         * A glob pattern, like `*.{ts,js}`.
-         */
-        pattern?: string;
-}
--}
-data DocumentFilter =
-  DocumentFilter
-    { _language :: Text
-    , _scheme   :: Text
-    , _pattern  :: Maybe Text
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''DocumentFilter
-makeFieldsNoPrefix ''DocumentFilter
-
-{-
-A document selector is the combination of one or many document filters.
-
-export type DocumentSelector = DocumentFilter[];
--}
-type DocumentSelector = List DocumentFilter
 
 -- =====================================================================
 -- ACTUAL PROTOCOL -----------------------------------------------------
@@ -1653,398 +1533,6 @@ type PublishDiagnosticsNotification = NotificationMessage ServerMethod PublishDi
 
 -- ---------------------------------------------------------------------
 {-
-Completion Request
-
-The Completion request is sent from the client to the server to compute
-completion items at a given cursor position. Completion items are presented in
-the IntelliSense user interface. If computing full completion items is
-expensive, servers can additionally provide a handler for the completion item
-resolve request ('completionItem/resolve'). This request is sent when a
-completion item is selected in the user interface. A typically use case is for
-example: the 'textDocument/completion' request doesn't fill in the documentation
-property for returned completion items since it is expensive to compute. When
-the item is selected in the user interface then a 'completionItem/resolve'
-request is sent with the selected completion item as a param. The returned
-completion item should have the documentation property filled in.
-
-    Changed: In 2.0 the request uses TextDocumentPositionParams with a proper
-    textDocument and position property. In 1.0 the uri of the referenced text
-    document was inlined into the params object.
-
-Request
-
-    method: 'textDocument/completion'
-    params: TextDocumentPositionParams
--}
-
--- -------------------------------------
-
-{-
-
-Response
-
-    result: CompletionItem[] | CompletionList
-
-/**
- * Represents a collection of [completion items](#CompletionItem) to be presented
- * in the editor.
- */
-interface CompletionList {
-    /**
-     * This list it not complete. Further typing should result in recomputing
-     * this list.
-     */
-    isIncomplete: boolean;
-    /**
-     * The completion items.
-     */
-    items: CompletionItem[];
-}
-
-
-New in 3.0 : InsertTextFormat
-
-/**
- * Defines whether the insert text in a completion item should be interpreted as
- * plain text or a snippet.
- */
-namespace InsertTextFormat {
-        /**
-         * The primary text to be inserted is treated as a plain string.
-         */
-        export const PlainText = 1;
-
-        /**
-         * The primary text to be inserted is treated as a snippet.
-         *
-         * A snippet can define tab stops and placeholders with `$1`, `$2`
-         * and `${3:foo}`. `$0` defines the final tab stop, it defaults to
-         * the end of the snippet. Placeholders with equal identifiers are linked,
-         * that is typing in one will update others too.
-         *
-         * See also: https://github.com/Microsoft/vscode/blob/master/src/vs/editor/contrib/snippet/common/snippet.md
-         */
-        export const Snippet = 2;
-}
-
-
-
-interface CompletionItem {
-    /**
-     * The label of this completion item. By default
-     * also the text that is inserted when selecting
-     * this completion.
-     */
-    label: string;
-    /**
-     * The kind of this completion item. Based of the kind
-     * an icon is chosen by the editor.
-     */
-    kind?: number;
-    /**
-     * A human-readable string with additional information
-     * about this item, like type or symbol information.
-     */
-    detail?: string;
-    /**
-     * A human-readable string that represents a doc-comment.
-     */
-    documentation?: string;
-    /**
-     * A string that shoud be used when comparing this item
-     * with other items. When `falsy` the label is used.
-     */
-    sortText?: string;
-    /**
-     * A string that should be used when filtering a set of
-     * completion items. When `falsy` the label is used.
-     */
-    filterText?: string;
-    /**
-     * A string that should be inserted a document when selecting
-     * this completion. When `falsy` the label is used.
-     */
-    insertText?: string;
-    -- Following field is new in 3.0
-        /**
-         * The format of the insert text. The format applies to both the `insertText` property
-         * and the `newText` property of a provided `textEdit`.
-         */
-    insertTextFormat?: InsertTextFormat;
-        /**
-         * An edit which is applied to a document when selecting this completion. When an edit is provided the value of
-         * `insertText` is ignored.
-         *
-         * *Note:* The range of the edit must be a single line range and it must contain the position at which completion
-         * has been requested.
-         */
-
-    textEdit?: TextEdit;
-
-    -- Following field is new in 3.0
-        /**
-         * An optional array of additional text edits that are applied when
-         * selecting this completion. Edits must not overlap with the main edit
-         * nor with themselves.
-         */
-    additionalTextEdits?: TextEdit[];
-    -- Following field is new in 3.0
-        /**
-         * An optional command that is executed *after* inserting this completion. *Note* that
-         * additional modifications to the current document should be described with the
-         * additionalTextEdits-property.
-         */
-
-    command?: Command;
-        /**
-         * An data entry field that is preserved on a completion item between
-         * a completion and a completion resolve request.
-         */
-
-    data?: any
-}
-
-Where CompletionItemKind is defined as follows:
-
-/**
- * The kind of a completion entry.
- */
-enum CompletionItemKind {
-    Text = 1,
-    Method = 2,
-    Function = 3,
-    Constructor = 4,
-    Field = 5,
-    Variable = 6,
-    Class = 7,
-    Interface = 8,
-    Module = 9,
-    Property = 10,
-    Unit = 11,
-    Value = 12,
-    Enum = 13,
-    Keyword = 14,
-    Snippet = 15,
-    Color = 16,
-    File = 17,
-    Reference = 18
-}
-
-    error: code and message set in case an exception happens during the completion request.
--}
-
--- -------------------------------------
-
-data InsertTextFormat
-  = PlainText -- ^The primary text to be inserted is treated as a plain string.
-  | Snippet
-      -- ^ The primary text to be inserted is treated as a snippet.
-      --
-      -- A snippet can define tab stops and placeholders with `$1`, `$2`
-      -- and `${3:foo}`. `$0` defines the final tab stop, it defaults to
-      -- the end of the snippet. Placeholders with equal identifiers are linked,
-      -- that is typing in one will update others too.
-      --
-      -- See also: https://github.com/Microsoft/vscode/blob/master/src/vs/editor/contrib/snippet/common/snippet.md
-    deriving (Show, Read, Eq)
-
-instance A.ToJSON InsertTextFormat where
-  toJSON PlainText = A.Number 1
-  toJSON Snippet   = A.Number 2
-
-instance A.FromJSON InsertTextFormat where
-  parseJSON (A.Number  1) = pure PlainText
-  parseJSON (A.Number  2) = pure Snippet
-  parseJSON _             = mempty
-
--- -------------------------------------
-
-data CompletionItemKind = CiText
-                        | CiMethod
-                        | CiFunction
-                        | CiConstructor
-                        | CiField
-                        | CiVariable
-                        | CiClass
-                        | CiInterface
-                        | CiModule
-                        | CiProperty
-                        | CiUnit
-                        | CiValue
-                        | CiEnum
-                        | CiKeyword
-                        | CiSnippet
-                        | CiColor
-                        | CiFile
-                        | CiReference
-         deriving (Read,Show,Eq,Ord)
-
-instance A.ToJSON CompletionItemKind where
-  toJSON CiText        = A.Number 1
-  toJSON CiMethod      = A.Number 2
-  toJSON CiFunction    = A.Number 3
-  toJSON CiConstructor = A.Number 4
-  toJSON CiField       = A.Number 5
-  toJSON CiVariable    = A.Number 6
-  toJSON CiClass       = A.Number 7
-  toJSON CiInterface   = A.Number 8
-  toJSON CiModule      = A.Number 9
-  toJSON CiProperty    = A.Number 10
-  toJSON CiUnit        = A.Number 11
-  toJSON CiValue       = A.Number 12
-  toJSON CiEnum        = A.Number 13
-  toJSON CiKeyword     = A.Number 14
-  toJSON CiSnippet     = A.Number 15
-  toJSON CiColor       = A.Number 16
-  toJSON CiFile        = A.Number 17
-  toJSON CiReference   = A.Number 18
-
-instance A.FromJSON CompletionItemKind where
-  parseJSON (A.Number  1) = pure CiText
-  parseJSON (A.Number  2) = pure CiMethod
-  parseJSON (A.Number  3) = pure CiFunction
-  parseJSON (A.Number  4) = pure CiConstructor
-  parseJSON (A.Number  5) = pure CiField
-  parseJSON (A.Number  6) = pure CiVariable
-  parseJSON (A.Number  7) = pure CiClass
-  parseJSON (A.Number  8) = pure CiInterface
-  parseJSON (A.Number  9) = pure CiModule
-  parseJSON (A.Number 10) = pure CiProperty
-  parseJSON (A.Number 11) = pure CiUnit
-  parseJSON (A.Number 12) = pure CiValue
-  parseJSON (A.Number 13) = pure CiEnum
-  parseJSON (A.Number 14) = pure CiKeyword
-  parseJSON (A.Number 15) = pure CiSnippet
-  parseJSON (A.Number 16) = pure CiColor
-  parseJSON (A.Number 17) = pure CiFile
-  parseJSON (A.Number 18) = pure CiReference
-  parseJSON _             = mempty
-
-
--- -------------------------------------
-
-data CompletionItem =
-  CompletionItem
-    { _label               :: Text -- ^ The label of this completion item. By default also
-                       -- the text that is inserted when selecting this
-                       -- completion.
-    , _kind                :: Maybe CompletionItemKind
-    , _detail              :: Maybe Text -- ^ A human-readable string with additional
-                              -- information about this item, like type or
-                              -- symbol information.
-    , _documentation       :: Maybe Text -- ^ A human-readable string that represents
-                                    -- a doc-comment.
-    , _sortText            :: Maybe Text -- ^ A string that should be used when filtering
-                                -- a set of completion items. When `falsy` the
-                                -- label is used.
-    , _filterText          :: Maybe Text -- ^ A string that should be used when
-                                  -- filtering a set of completion items. When
-                                  -- `falsy` the label is used.
-    , _insertText          :: Maybe Text -- ^ A string that should be inserted a
-                                  -- document when selecting this completion.
-                                  -- When `falsy` the label is used.
-    , _insertTextFormat    :: Maybe InsertTextFormat
-         -- ^ The format of the insert text. The format applies to both the
-         -- `insertText` property and the `newText` property of a provided
-         -- `textEdit`.
-    , _textEdit            :: Maybe TextEdit
-         -- ^ An edit which is applied to a document when selecting this
-         -- completion. When an edit is provided the value of `insertText` is
-         -- ignored.
-         --
-         -- *Note:* The range of the edit must be a single line range and it
-         -- must contain the position at which completion has been requested.
-    , _additionalTextEdits :: Maybe (List TextEdit)
-         -- ^ An optional array of additional text edits that are applied when
-         -- selecting this completion. Edits must not overlap with the main edit
-         -- nor with themselves.
-    , _command             :: Maybe Command
-        -- ^ An optional command that is executed *after* inserting this
-        -- completion. *Note* that additional modifications to the current
-        -- document should be described with the additionalTextEdits-property.
-    , _xdata               :: Maybe A.Value -- ^ An data entry field that is preserved on a
-                              -- completion item between a completion and a
-                              -- completion resolve request.
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions{ fieldLabelModifier = customModifier } ''CompletionItem
-makeFieldsNoPrefix ''CompletionItem
-
-data CompletionListType =
-  CompletionListType
-    { _isIncomplete :: Bool
-    , _items        :: List CompletionItem
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''CompletionListType
-makeFieldsNoPrefix ''CompletionListType
-
-
-data CompletionResponseResult
-  = CompletionList CompletionListType
-  | Completions (List CompletionItem)
-  deriving (Read,Show,Eq)
-
-deriveJSON defaultOptions { fieldLabelModifier = rdrop (length ("CompletionResponseResult"::String)), sumEncoding = UntaggedValue } ''CompletionResponseResult
-
-type CompletionResponse = ResponseMessage CompletionResponseResult
-type CompletionRequest = RequestMessage ClientMethod TextDocumentPositionParams CompletionResponseResult
-
--- -------------------------------------
-{-
-New in 3.0
------------
-Registration Options: CompletionRegistrationOptions options defined as follows:
-
-export interface CompletionRegistrationOptions extends TextDocumentRegistrationOptions {
-        /**
-         * The characters that trigger completion automatically.
-         */
-        triggerCharacters?: string[];
-
-        /**
-         * The server provides support to resolve additional
-         * information for a completion item.
-         */
-        resolveProvider?: boolean;
-}
--}
-
-data CompletionRegistrationOptions =
-  CompletionRegistrationOptions
-    { _documentSelector  :: Maybe DocumentSelector
-    , _triggerCharacters :: Maybe (List String)
-    , _resolveProvider   :: Maybe Bool
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''CompletionRegistrationOptions
-makeFieldsNoPrefix ''CompletionRegistrationOptions
-
--- ---------------------------------------------------------------------
-{-
-Completion Item Resolve Request
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#completion-item-resolve-request
-
-The request is sent from the client to the server to resolve additional
-information for a given completion item.
-
-Request
-
-    method: 'completionItem/resolve'
-    params: CompletionItem
-
-Response
-
-    result: CompletionItem
-    error: code and message set in case an exception happens during the completion resolve request.
--}
-
-type CompletionItemResolveRequest  = RequestMessage ClientMethod CompletionItem CompletionItem
-type CompletionItemResolveResponse = ResponseMessage CompletionItem
-
--- ---------------------------------------------------------------------
-{-
 Hover Request
 
 The hover request is sent from the client to the server to request hover
@@ -3201,10 +2689,18 @@ makeFieldsNoPrefix ''Position
 makeFieldsNoPrefix ''Range
 makeFieldsNoPrefix ''Location
 
+-- Completion
+makeFieldsNoPrefix ''CompletionItem
+makeFieldsNoPrefix ''CompletionListType
+makeFieldsNoPrefix ''CompletionRegistrationOptions
+
 -- CodeActions
 makeFieldsNoPrefix ''CodeActionContext
 makeFieldsNoPrefix ''CodeActionParams
 makeFieldsNoPrefix ''CodeAction
+
+-- DocumentFilter
+makeFieldsNoPrefix ''DocumentFilter
 
 -- WorkspaceEdit
 makeFieldsNoPrefix ''TextEdit
@@ -3219,8 +2715,10 @@ makeFieldsNoPrefix ''ResponseMessage
 makeFieldsNoPrefix ''NotificationMessage
 makeFieldsNoPrefix ''CancelParams
 
--- TextDocumentIdentifier
+-- TextDocument
+makeFieldsNoPrefix ''TextDocumentItem
 makeFieldsNoPrefix ''TextDocumentIdentifier
+makeFieldsNoPrefix ''TextDocumentPositionParams
 
 -- Command
 makeFieldsNoPrefix ''Command
