+++ haskell-lsp-types/src/Language/Haskell/LSP/TH/Symbol.hs
@@ -5,7 +5,7 @@ import           Data.Aeson
 import           Data.Aeson.TH
 import           Data.Text                                      (Text)
 import           Language.Haskell.LSP.TH.Constants
-import           Language.Haskell.LSP.TH.TextDocumentIdentifier
+import           Language.Haskell.LSP.TH.TextDocument
 import           Language.Haskell.LSP.TH.List
 import           Language.Haskell.LSP.TH.Location
 import           Language.Haskell.LSP.TH.Message
